Initial layout classes refactor
