+++ library/include/rocwmma/internal/layout/layout_profile.hpp
@@ -23,132 +23,27 @@
  * SOFTWARE.
  *
  *******************************************************************************/
-#ifndef ROCWMMA_LAYOUT_HPP
-#define ROCWMMA_LAYOUT_HPP
+#ifndef ROCWMMA_LAYOUT_PROFILE_HPP
+#define ROCWMMA_LAYOUT_PROFILE_HPP
 
-#include "mapping_util.hpp"
+#include "layout.hpp"
 
 namespace rocwmma
 {
-    // In relation to matrix space, DataLayouts describe whether consecutive elements in 1D data arrays are:
-    // 1. Contiguous rows (row_major)
-    // 2. Contiguous columns (col_major)
-    namespace DataLayout
-    {
-        template <typename DataLayoutT>
-        using Array1d = typename ::rocwmma::detail::template DataSpace<DataLayoutT>;
-
-        using RowMajor = Array1d<row_major>;
-        using ColMajor = Array1d<col_major>;
-
-    } // namespace DataLayout
-
-    // In 2D space, Matrix Layouts describe per-thread offset coordinates and iterative spaces
-    // 1. Base thread offsets
-    // 2. Stride offsets
-    // 3. Stride spaces (counts)
-    // 4. Per-iteration offsets (stride step based on iteration)
-    // 5. Cumulative offsets (cumulative stride steps based on iteration)
-    namespace MatrixLayout
-    {
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth>
-        struct ColOrthoVW;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth>
-        struct ColInlineVW;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth>
-        struct RowOrthoVW;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth>
-        struct RowInlineVW;
-
-        /////////////////// Interleaved patterns //////////////////
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth, //
-                  uint32_t MfmaDim, // MFMA instruction size
-                  uint32_t SplitK = 1> // # of splits
-        struct ColInlineInt;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth, //
-                  uint32_t MfmaDim, // MFMA instruction size
-                  uint32_t SplitK = 1> // # of splits
-        struct ColOrthoInt;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth, //
-                  uint32_t MfmaDim, // MFMA instruction size
-                  uint32_t SplitK = 1> // # of splits
-        struct RowInlineInt;
-
-        template <uint32_t BlockDim,
-                  uint32_t BlockK,
-                  typename DataT,
-                  uint32_t VectorWidth,
-                  uint32_t MaxVectorWidth, //
-                  uint32_t MfmaDim, // MFMA instruction size
-                  uint32_t SplitK = 1> // # of splits
-        struct RowOrthoInt;
-
-        /////////////////// ////////////////////////////// //////////////////
-
-    } // namespace MatrixLayout
-
-    // Register layouts describe whether contiguous BlockDim elements are:
-    // 1. Captured in the same register lane as if the input were in Array-Of-Structures (AOS)
-    // 2. Captured across multiple register lanes as if the input were in Structure-Of-Arrays (SOA)
-    namespace RegisterLayout
-    {
-        template <uint32_t BlockDim, uint32_t VW>
-        struct Aos
-        {
-        };
-        template <uint32_t BlockDim, uint32_t VW>
-        struct Soa
-        {
-        };
-    }
-
-    // Layout profiles describe fragments in three mapped spaces:
-    // 1. DataLayout:     data locality in memory space (row_major or col_major)
-    // 2. MatrixLayout:   data locality in matrix space (ColOrthoVW, ColInlineVW, etc.)
-    // 3. RegisterLayout: data locality in register space (AOS or SOA)
+    // Layout profiles are high-level objects that describe fragments in three mapped spaces:
+    // 1. DataLayout:     data locality in 1D memory space (row_major or col_major)
+    // 2. MatrixLayout:   data locality in 2D matrix space (ColOrthoVW, ColInlineVW, etc.)
+    // 3. RegisterLayout: data locality in register space (Storage, or MmaInput)
     namespace LayoutProfile
     {
         // ColNT is a layout profile that has the following properties:
         // 1. Leading dimension is aligned with column elements of fragment data:
         //    - BlockDim is assumed the column size, or BlockM dimension.
         //    - Analogous to capturing columns of 'matrix A'.
-        // 2. Register elements are in MFMA friendly, or SOA register layout.
-        // 3. Register layout does NOT change whether DataLayout is col_major or row_major (fast DataLayoutT change).
-        // 4. MatrixLayout will capture contiguous column elements across multiple register lanes.
-        // 5. VectorWidth is fixed to 1 in col_major to ensure #3 (non-optimal).
+        // 2. When BlockDim is supported by mma, register elements are always in MmaInput friendly register layout.
+        // 3. Register layout does NOT change whether DataLayout is col_major or row_major (free DataLayoutT change).
+        // 4. MatrixLayout will capture contiguous column elements across contiguous threads.
+        // 5. VectorWidth is fixed to 1 in col_major to ensure #4 (non-optimal).
         template <uint32_t BlockDim,
                   uint32_t BlockK,
                   typename DataT,
@@ -158,12 +53,14 @@ namespace rocwmma
         struct ColNT
         {
             // Layouts
-            using DataLayout   = DataLayout::template Array1d<DataLayoutT>;
+            using DataLayout = DataLayout::template Array1d<DataLayoutT>;
+
             using MatrixLayout = conditional_t<
                 is_same_v<DataLayoutT, col_major>,
                 MatrixLayout::ColOrthoVW<BlockDim, BlockK, DataT, 1, MaxVectorWidth>,
-                MatrixLayout::ColOrthoVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>>;
-            using RegisterLayout = RegisterLayout::template Soa<BlockDim, MaxVectorWidth>;
+                MatrixLayout::ColOrthoVW<BlockDim, BlockK, DataT, MaxVectorWidth, MaxVectorWidth>>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -187,10 +84,10 @@ namespace rocwmma
         // 1. Leading dimension is aligned with row elements of fragment data:
         //    - BlockDim is assumed the row size, or BlockN dimension.
         //    - Analogous to capturing rows of 'matrix B' or 'accumulator'.
-        // 2. Register elements are in MFMA friendly, or SOA register layout.
+        // 2. When BlockDim is supported by mma, register elements are always MmaInput friendly register layout.
         // 3. Register layout does NOT change whether DataLayout is col_major or row_major (fast DataLayoutT change).
-        // 4. MatrixLayout will capture contiguous row elements across multiple register lanes.
-        // 5. VectorWidth is fixed to 1 in row_major to ensure #3 (non-optimal).
+        // 4. MatrixLayout will capture contiguous row elements across contiguous threads.
+        // 5. VectorWidth is fixed to 1 in row_major to ensure #4 (non-optimal).
         template <uint32_t BlockDim,
                   uint32_t BlockK,
                   typename DataT,
@@ -200,12 +97,14 @@ namespace rocwmma
         struct RowNT
         {
             // Layouts
-            using DataLayout   = DataLayout::template Array1d<DataLayoutT>;
+            using DataLayout = DataLayout::template Array1d<DataLayoutT>;
+
             using MatrixLayout = conditional_t<
                 is_same_v<DataLayoutT, col_major>,
                 MatrixLayout::RowOrthoVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>,
                 MatrixLayout::RowOrthoVW<BlockDim, BlockK, DataT, 1, MaxVectorWidth>>;
-            using RegisterLayout = RegisterLayout::template Soa<BlockDim, MaxVectorWidth>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -229,12 +128,11 @@ namespace rocwmma
         //    - BlockDim is assumed the column size, or BlockM dimension.
         //    - Analogous to capturing columns of 'matrix A'.
         // 2. Register layout is dynamic:
-        //    - col_major data is stored in AOS register layout (non-MFMA friendly), and
-        //    - row_major data is stored in SOA register layout (MFMA friendly).
-        //    - Both data layouts cover the same geometric elements (transform friendly).
+        //    - col_major data is stored in AOS register layout (non-MmaInput friendly), and
+        //    - row_major data is stored in SOA register layout (MmaInput friendly).
         // 3. Register layout DOES change whether DataLayout is col_major or row_major (cost for DataLayoutT change).
         // 4. VectorWidth is NOT fixed to 1 in either data layout (optimal).
-        // 5. User must convert to SOA if using with MFMA.
+        // 5. Must convert to SOA if using with MFMA.
         template <uint32_t BlockDim,
                   uint32_t BlockK,
                   typename DataT,
@@ -244,15 +142,14 @@ namespace rocwmma
         struct Col
         {
             // Layouts
-            using DataLayout   = DataLayout::template Array1d<DataLayoutT>;
+            using DataLayout = DataLayout::template Array1d<DataLayoutT>;
+
             using MatrixLayout = conditional_t<
                 is_same_v<DataLayoutT, col_major>,
                 MatrixLayout::ColInlineVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>,
                 MatrixLayout::ColOrthoVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>>;
-            using RegisterLayout
-                = conditional_t<is_same_v<DataLayoutT, col_major>,
-                                RegisterLayout::template Aos<BlockDim, MaxVectorWidth>,
-                                RegisterLayout::template Soa<BlockDim, MaxVectorWidth>>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -290,10 +187,8 @@ namespace rocwmma
                 is_same_v<DataLayoutT, row_major>,
                 MatrixLayout::RowInlineVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>,
                 MatrixLayout::RowOrthoVW<BlockDim, BlockK, DataT, VectorWidth, MaxVectorWidth>>;
-            using RegisterLayout
-                = conditional_t<is_same_v<DataLayoutT, row_major>,
-                                RegisterLayout::template Aos<BlockDim, MaxVectorWidth>,
-                                RegisterLayout::template Soa<BlockDim, MaxVectorWidth>>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -346,10 +241,8 @@ namespace rocwmma
                                                                          MaxVectorWidth,
                                                                          MfmaDim,
                                                                          SplitK>>;
-            using RegisterLayout
-                = conditional_t<is_same_v<DataLayoutT, col_major>,
-                                RegisterLayout::template Aos<BlockDim, MaxVectorWidth>,
-                                RegisterLayout::template Soa<BlockDim, MaxVectorWidth>>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -400,10 +293,8 @@ namespace rocwmma
                                                                          MaxVectorWidth,
                                                                          MfmaDim,
                                                                          SplitK>>;
-            using RegisterLayout
-                = conditional_t<is_same_v<DataLayoutT, row_major>,
-                                RegisterLayout::template Aos<BlockDim, MaxVectorWidth>,
-                                RegisterLayout::template Soa<BlockDim, MaxVectorWidth>>;
+
+            using RegisterLayout = RegisterLayout::Storage<MatrixLayout>;
 
             // Mapping
             using MappingUtil  = MappingUtil<BlockDim, BlockK, DataT, DataLayoutT>;
@@ -416,38 +307,8 @@ namespace rocwmma
                 "MaxVectorWidth is larger than BlockK dimension. Try reducing MaxVectorWidth");
         };
 
-    } // namespace FragmentLayout
-
-    ///
-    /// Helper to ensure layout types are consistent (same, or equivalent)
-    ///
-    template <typename LhsLayout, typename RhsLayout>
-    struct ConsistencyCheck : public false_type
-    {
-    };
-
-    ///
-    /// Check for layout orthogonality
-    ///
-    template <typename LhsMatrixLayout, typename RhsMatrixLayout>
-    struct OrthogonalCheck : public false_type
-    {
-    };
-
-    template <typename LayoutT>
-    struct OrthogonalLayout;
-
-    template <typename Layout>
-    using orthogonal_layout_t = typename OrthogonalLayout<Layout>::Type;
-
-    template <typename LhsDataLayout, typename RhsDataLayout>
-    struct is_orthogonal;
-
-    template <typename LhsDataLayout, typename RhsDataLayout>
-    inline constexpr bool is_orthogonal_v = is_orthogonal<LhsDataLayout, RhsDataLayout>::value;
+    } // namespace LayoutProfile
 
 } // namespace rocwmma
 
-#include "layout_impl.hpp"
-
-#endif // ROCWMMA_LAYOUT_HPP
+#endif // ROCWMMA_LAYOUT_PROFILE_HPP
