+++ library/include/rocwmma/internal/layout/register_layout_impl.hpp
@@ -1,0 +1,44 @@
+/*******************************************************************************
+ *
+ * MIT License
+ *
+ * Copyright (C) 2021-2024 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ *******************************************************************************/
+#ifndef ROCWMMA_REGISTER_LAYOUT_IMPL_HPP
+#define ROCWMMA_REGISTER_LAYOUT_IMPL_HPP
+
+#include "layout.hpp"
+#include "layout_traits.hpp"
+#include "utility/type_traits.hpp"
+
+namespace rocwmma
+{
+    // Use generic MatrixLayout transpose rules to guide the register layout transpose suggestion
+    template <typename MatrixLayout>
+    struct layout_transpose<RegisterLayout::template Storage<MatrixLayout>>
+    {
+        using type = RegisterLayout::template Storage<layout_transpose_t<MatrixLayout>>;
+    };
+
+} // namespace rocwmma
+
+#endif // ROCWMMA_REGISTER_LAYOUT_IMPL_HPP
